move Projects continent components into their own Projects folder
