+++ src/components/app/App.jsx
@@ -6,7 +6,7 @@ import {
 } from 'react-router-dom';
 import Ocean from '../ocean/Ocean';
 import ProfileDetails from '../continents/Profile/ProfileDetails';
-import ProjectsDetails from '../continents/ProjectsDetails';
+import ProjectsDetails from '../continents/Projects/ProjectsDetails';
 import BeHumanDetails from '../projects/BeHuman/BeHumanDetails';
 import ChooseGifDetails from '../projects/ChooseGif/ChooseGifDetails';
 import WeatherBeatsDetails from '../projects/WeatherBeats/WeatherBeatsDetails';
