+++ src/components/continents/Projects/ProjectsDetails.module.css
