+++ src/components/ocean/Ocean.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 // import Continent from '../continents/Continent';
 // import { continentData } from '../continents/continentData';
 // import resume from '../../assets/FrancoOrtega_Resume.pdf';
-import Projects from '../continents/Projects';
-import Profile from '../continents//Profile/Profile';
+import Projects from '../continents/Projects/Projects';
+import Profile from '../continents/Profile/Profile';
 import styles from './Ocean.module.css';
 import Resume from '../continents/Resume';
 import Scale from '../mapElements/Scale';
